Refactor consent calls
