+++ clients/privacy-center/components/consent/ConsentToggles.tsx
@@ -1,13 +1,19 @@
 import { useSettings } from "~/features/common/settings.slice";
+import { ConsentPreferences } from "~/types/api";
+
 import ConfigDrivenConsent from "./ConfigDrivenConsent";
 import NoticeDrivenConsent from "./NoticeDrivenConsent";
 
-const ConsentToggles = () => {
+const ConsentToggles = ({
+  storePreferences,
+}: {
+  storePreferences: (data: ConsentPreferences) => void;
+}) => {
   const settings = useSettings();
   const { IS_OVERLAY_DISABLED } = settings;
 
   if (IS_OVERLAY_DISABLED) {
-    return <ConfigDrivenConsent />;
+    return <ConfigDrivenConsent storePreferences={storePreferences} />;
   }
   return <NoticeDrivenConsent />;
 };
